add test DOT from relay to para through asset hub
